fix(ui): record tool outcomes and repair transcript row regressions
Neutral collapsed rows are the design, but this branch left failures with
nowhere to land and shipped chrome whose styles it had already deleted.

- Expanded standalone tool messages close with the outcome line, so an inline
  rows shape still reports the failure instead of dropping it.
- Progress-card receipts drop the badge and the spinner: this branch removed
  both styles, leaving a 0x0 invisible running indicator, and the receipt
  label already names the state.
- Single delete patches are non-navigable; their workspace target is gone, so
  the file action could only ever fail to load.
- Flush command padding derives from the action-button size instead of a stale
  42px constant, and the running text wave covers command rows again.
- Activity lane returns to the tool-shell column width the rule claims.
- Drop CSS nothing references: tool-card tabs, empty, plain-output, activity
  scroll masks, secondary command text.
- renderToolDataBlock takes an optional label instead of comparing copy.
- Split the oversized tool-card test file and refresh stale row assertions.
