+++ ui/src/e2e/chat-tool-turn-outcome.e2e.test.ts
@@ -92,18 +92,18 @@ suite.define(() => {
       "Tool output",
       "Tool output",
     ]);
-    // Each failure keeps only its per-call badge even when its turn later
-    // recovers; both row summaries otherwise render neutral.
+    // Collapsed rows stay neutral even when the call failed; the failure is
+    // recorded as the expanded body's outcome, with the reported exit code.
     const summaryClasses = await page
       .locator(".chat-tool-msg-summary")
       .evaluateAll((nodes) => nodes.map((node) => node.className));
     expect(summaryClasses).toHaveLength(2);
     expect(summaryClasses[0]).not.toContain("chat-tool-msg-summary--error");
     expect(summaryClasses[1]).not.toContain("chat-tool-msg-summary--error");
-    expect(await page.locator(".chat-tool-row__badge").allTextContents()).toEqual([
-      "failed",
-      "failed",
-    ]);
+    await page.locator(".chat-tool-msg-summary").first().click();
+    await expect
+      .poll(() => page.locator(".chat-tool-card__outcome").first().textContent())
+      .toBe("Exit code 1");
     await context.close();
   });
 
@@ -175,7 +175,7 @@ suite.define(() => {
     const activityGeometry = await activity.evaluate((node) => {
       const container = node.closest<HTMLElement>(".chat-activity-group");
       const label = node.querySelector<HTMLElement>(".chat-activity-group__label");
-      const chevron = node.querySelector<HTMLElement>(".chat-inline-disclosure__chevron");
+      const chevron = node.querySelector<HTMLElement>(".chat-tool-row__chevron");
       if (!container || !label || !chevron) {
         throw new Error("Expected compact activity disclosure parts");
       }
@@ -201,11 +201,13 @@ suite.define(() => {
 
     const rows = page.locator(".chat-activity-group__body .chat-tool-msg-summary");
     expect(await rows.count()).toBe(2);
-    expect(await rows.locator(".chat-inline-disclosure__chevron").count()).toBe(2);
+    expect(await rows.locator(".chat-tool-row__chevron").count()).toBe(2);
     expect(await page.locator(".chat-tool-msg-summary__label", { hasText: "Tool" }).count()).toBe(
       0,
     );
-    await rows.first().click();
+    // File rows put the workspace link inside the row, so toggle from the icon
+    // edge instead of the row centre to avoid opening the linked file.
+    await rows.first().click({ position: { x: 4, y: 4 } });
     expect(await page.getByText("offset:", { exact: true }).count()).toBe(1);
     expect(await page.getByText("limit:", { exact: true }).count()).toBe(1);
     const patchRow = rows.filter({ hasText: "2 files" });
@@ -385,7 +387,8 @@ suite.define(() => {
     await card.waitFor();
     expect(await card.getByText("query:", { exact: true }).count()).toBe(1);
     expect(await card.getByText("example", { exact: true }).count()).toBe(1);
-    await card.getByText("Tool output", { exact: true }).waitFor();
+    // Plain output needs no "Tool output" header; the payload is the content.
+    expect(await card.getByText("Tool output", { exact: true }).count()).toBe(0);
     await card.getByText("Native result payload", { exact: true }).waitFor();
     await captureToolActivityProof(page, "native-result-before-call-expanded");
     await context.close();
